Improve importCoin API
